Crate aarch64_sysreg

Aarch64 System Register

This crate provides a set of types and operations for working with low-level system registers.

Features

OperationType: Defines different types of operations that can be performed.

RegistersType: Defines various types of registers.

SystemRegType: Defines specific system registers.

Enums

OperationType

Arm64 Operation type

RegistersType

Register type

SystemRegType

System register type